requirements.txt
astroid==4.0.3
certifi==2026.1.4
charset-normalizer==3.4.4
dill==0.4.1
docker==7.1.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
isort==7.0.0
mccabe==0.7.0
platformdirs==4.5.1
pylint==4.0.4
requests==2.32.5
smmap==5.0.2
tomlkit==0.14.0
urllib3==2.6.3
